Update index.md
Converted the topic to a conceptual overview instead of something that feels like release notes. The Identity Engine has been out for 18 months, so I tried to focus the topic more on concepts.

Removed “overview” from the main heading as this is rarely used in the doc set.

Created a bulleted list of the new features for scanning (didn’t put in-topic links here on purpose so as not to overwhelm with link text)

Reused a lot of the conceptual information from other Identity Engine topics to add more “meat” to this topic. 

Removed the detailed description about App intent links used to work. I would talk to a SME about this as there is very little doc support for App intent links that I could find. Arguable to leave it in, but it seemed overkill for an overview topic.

Minor editing to more user-focused writing when appropriate. For example, “You can do this” instead of “Identity Engine does this.”

Moved upgrade information to a note and put it higher up in the topic.

Deleted the topic that documented the documentation. I think the tagging strategy for Identity Engine is quite intuitive and the Classic Engine notes are obvious as well.